Update streamlit_app.py
Remove 'European Union - 27 countries (from 2020)' from the default country selection. As this is the sum of most countries, it has raised the scale on both graphs to the point where the rest are indistinguishable.

Set the default map date to the date of the first record. Necessary because in most cases the most recent record is missing data, as countries tend to be late in reporting to the EU.